+++ common/src/main/java/com/chefmooon/ubesdelight/common/block/leaf_feast/LumpiaLeafFeastBlock.java
@@ -6,7 +6,7 @@ import com.chefmooon.ubesdelight.common.registry.UbesDelightShapes;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.phys.shapes.VoxelShape;
 
-import java.util.HashMap;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.function.Supplier;
 
 public class LumpiaLeafFeastBlock extends LargeLeafFeastBlock {
@@ -14,26 +14,15 @@ public class LumpiaLeafFeastBlock extends LargeLeafFeastBlock {
         super(servingItem, properties, buildShapes());
     }
 
-    public static HashMap<LeafFeastTypes, HashMap<Integer, VoxelShape>> buildShapes() {
-        HashMap<LeafFeastTypes, HashMap<Integer, VoxelShape>> result = new HashMap<>();
+    public static ConcurrentHashMap<LeafFeastTypes, ConcurrentHashMap<Integer, VoxelShape>> buildShapes() {
+        ConcurrentHashMap<LeafFeastTypes, ConcurrentHashMap<Integer, VoxelShape>> result = new ConcurrentHashMap<>();
 
-        HashMap<Integer, VoxelShape> baseServingShapes = new HashMap<>();
-        baseServingShapes.put(1, UbesDelightShapes.LUMPIA_SERVING_1_OF_3);
-        baseServingShapes.put(2, UbesDelightShapes.LUMPIA_SERVING_2_OF_3);
-        baseServingShapes.put(3, UbesDelightShapes.LUMPIA_SERVING_3_OF_3);
+        ConcurrentHashMap<Integer, VoxelShape> middleServingShapes = new ConcurrentHashMap<>();
+        middleServingShapes.put(1, UbesDelightShapes.LUMPIA_SERVING_1_OF_3);
+        middleServingShapes.put(2, UbesDelightShapes.LUMPIA_SERVING_2_OF_3);
+        middleServingShapes.put(3, UbesDelightShapes.LUMPIA_SERVING_3_OF_3);
 
-//        HashMap<Integer, VoxelShape> tipServingShapes = new HashMap<>();
-//        tipServingShapes.put(1, UbesDelightShapes.PANDESAL_TIP_SERVING_1);
-//        tipServingShapes.put(2, UbesDelightShapes.PANDESAL_TIP_SERVING_2);
-//        tipServingShapes.put(3, UbesDelightShapes.PANDESAL_TIP_SERVING_3);
-//
-//        HashMap<Integer, VoxelShape> endServingShapes = new HashMap<>();
-//        endServingShapes.put(1, UbesDelightShapes.PANDESAL_END_SERVING_1);
-//        endServingShapes.put(2, UbesDelightShapes.PANDESAL_END_SERVING_2);
-//        endServingShapes.put(3, UbesDelightShapes.PANDESAL_END_SERVING_3);
-
-//        result.put(LeafFeastTypes.BASE, baseServingShapes);
-        result.put(LeafFeastTypes.MIDDLE, baseServingShapes);
+        result.put(LeafFeastTypes.MIDDLE, middleServingShapes);
 
         return result;
     }