+++ fabric/src/main/java/com/chefmooon/ubesdelight/common/registry/fabric/UbesDelightItemsImpl.java
@@ -63,21 +63,38 @@ public class UbesDelightItemsImpl {
 
     public static final Item LEAF_FEAST = registerItemWithTab(UbesDelightItems.LEAF_FEAST,
             new BlockItem(UbesDelightBlocksImpl.LEAF_FEAST, basicItem()));
-
     public static final Item LUMPIA_FEAST = registerItemWithTab(UbesDelightItems.LUMPIA_FEAST,
-            new BlockItem(UbesDelightBlocksImpl.LUMPIA_FEAST, noStack()));
+            new BlockItem(UbesDelightBlocksImpl.LUMPIA_FEAST, halfLeafFeast()));
+
     public static final Item LEAF_FEAST_ENSAYMADA = registerItemWithTab(UbesDelightItems.LEAF_FEAST_ENSAYMADA,
-            new BlockItem(UbesDelightBlocksImpl.LEAF_FEAST_ENSAYMADA, noStack()));
+            new BlockItem(UbesDelightBlocksImpl.LEAF_FEAST_ENSAYMADA, leafFeast()));
+    public static final Item LEAF_FEAST_ENSAYMADA_HALF = registerItemWithTab(UbesDelightItems.LEAF_FEAST_ENSAYMADA_HALF,
+            new BreadLeafFeastItem(UbesDelightBlocksImpl.LEAF_FEAST_ENSAYMADA, halfLeafFeast().component(UbesDelightDataComponentTypesImpl.SIMPLE_LEAF_FEAST_SERVINGS, 3)));
+
     public static final Item LEAF_FEAST_ENSAYMADA_UBE = registerItemWithTab(UbesDelightItems.LEAF_FEAST_ENSAYMADA_UBE,
-            new BlockItem(UbesDelightBlocksImpl.LEAF_FEAST_ENSAYMADA_UBE, noStack()));
+            new BlockItem(UbesDelightBlocksImpl.LEAF_FEAST_ENSAYMADA_UBE, leafFeast()));
+    public static final Item LEAF_FEAST_ENSAYMADA_UBE_HALF = registerItemWithTab(UbesDelightItems.LEAF_FEAST_ENSAYMADA_UBE_HALF,
+            new BreadLeafFeastItem(UbesDelightBlocksImpl.LEAF_FEAST_ENSAYMADA_UBE, halfLeafFeast().component(UbesDelightDataComponentTypesImpl.SIMPLE_LEAF_FEAST_SERVINGS, 3)));
+
     public static final Item LEAF_FEAST_PANDESAL = registerItemWithTab(UbesDelightItems.LEAF_FEAST_PANDESAL,
-            new BlockItem(UbesDelightBlocksImpl.LEAF_FEAST_PANDESAL, noStack()));
+            new BlockItem(UbesDelightBlocksImpl.LEAF_FEAST_PANDESAL, leafFeast()));
+    public static final Item LEAF_FEAST_PANDESAL_HALF = registerItemWithTab(UbesDelightItems.LEAF_FEAST_PANDESAL_HALF,
+            new BreadLeafFeastItem(UbesDelightBlocksImpl.LEAF_FEAST_PANDESAL, halfLeafFeast().component(UbesDelightDataComponentTypesImpl.SIMPLE_LEAF_FEAST_SERVINGS, 3)));
+
     public static final Item LEAF_FEAST_PANDESAL_UBE = registerItemWithTab(UbesDelightItems.LEAF_FEAST_PANDESAL_UBE,
-            new BlockItem(UbesDelightBlocksImpl.LEAF_FEAST_PANDESAL_UBE, noStack()));
+            new BlockItem(UbesDelightBlocksImpl.LEAF_FEAST_PANDESAL_UBE, leafFeast()));
+    public static final Item LEAF_FEAST_PANDESAL_UBE_HALF = registerItemWithTab(UbesDelightItems.LEAF_FEAST_PANDESAL_UBE_HALF,
+            new BreadLeafFeastItem(UbesDelightBlocksImpl.LEAF_FEAST_PANDESAL_UBE, halfLeafFeast().component(UbesDelightDataComponentTypesImpl.SIMPLE_LEAF_FEAST_SERVINGS, 3)));
+
     public static final Item LEAF_FEAST_HOPIA_MUNGGO = registerItemWithTab(UbesDelightItems.LEAF_FEAST_HOPIA_MUNGGO,
-            new BlockItem(UbesDelightBlocksImpl.LEAF_FEAST_HOPIA_MUNGGO, noStack()));
+            new BlockItem(UbesDelightBlocksImpl.LEAF_FEAST_HOPIA_MUNGGO, leafFeast()));
+    public static final Item LEAF_FEAST_HOPIA_MUNGGO_HALF = registerItemWithTab(UbesDelightItems.LEAF_FEAST_HOPIA_MUNGGO_HALF,
+            new BreadLeafFeastItem(UbesDelightBlocksImpl.LEAF_FEAST_HOPIA_MUNGGO, halfLeafFeast().component(UbesDelightDataComponentTypesImpl.SIMPLE_LEAF_FEAST_SERVINGS, 3)));
+
     public static final Item LEAF_FEAST_HOPIA_UBE = registerItemWithTab(UbesDelightItems.LEAF_FEAST_HOPIA_UBE,
-            new BlockItem(UbesDelightBlocksImpl.LEAF_FEAST_HOPIA_UBE, noStack()));
+            new BlockItem(UbesDelightBlocksImpl.LEAF_FEAST_HOPIA_UBE, leafFeast()));
+    public static final Item LEAF_FEAST_HOPIA_UBE_HALF = registerItemWithTab(UbesDelightItems.LEAF_FEAST_HOPIA_UBE_HALF,
+            new BreadLeafFeastItem(UbesDelightBlocksImpl.LEAF_FEAST_HOPIA_UBE, halfLeafFeast().component(UbesDelightDataComponentTypesImpl.SIMPLE_LEAF_FEAST_SERVINGS, 3)));
 
     public static final Item POISONOUS_UBE = registerItemWithTab(UbesDelightItems.POISONOUS_UBE,
             new Item(basicItem()));