+++ neoforge/src/main/java/com/chefmooon/ubesdelight/common/registry/neoforge/UbesDelightItemsImpl.java
@@ -72,19 +72,37 @@ public class UbesDelightItemsImpl {
             () -> new BlockItem(UbesDelightBlocksImpl.LEAF_FEAST.get(), basicItem()));
 
     public static final Supplier<Item> LUMPIA_FEAST = registerItemWithTab(UbesDelightItems.LUMPIA_FEAST,
-            () -> new BlockItem(UbesDelightBlocksImpl.LUMPIA_FEAST.get(), noStack()));
+            () -> new BlockItem(UbesDelightBlocksImpl.LUMPIA_FEAST.get(), halfLeafFeast()));
     public static final Supplier<Item> LEAF_FEAST_ENSAYMADA = registerItemWithTab(UbesDelightItems.LEAF_FEAST_ENSAYMADA,
-            () -> new BlockItem(UbesDelightBlocksImpl.LEAF_FEAST_ENSAYMADA.get(), noStack()));
+            () -> new BlockItem(UbesDelightBlocksImpl.LEAF_FEAST_ENSAYMADA.get(), leafFeast()));
+    public static final Supplier<Item> LEAF_FEAST_ENSAYMADA_HALF = registerItemWithTab(UbesDelightItems.LEAF_FEAST_ENSAYMADA_HALF,
+            () -> new BreadLeafFeastItem(UbesDelightBlocksImpl.LEAF_FEAST_ENSAYMADA.get(), halfLeafFeast().component(UbesDelightDataComponentTypesImpl.SIMPLE_LEAF_FEAST_SERVINGS, 3)));
+
     public static final Supplier<Item> LEAF_FEAST_ENSAYMADA_UBE = registerItemWithTab(UbesDelightItems.LEAF_FEAST_ENSAYMADA_UBE,
-            () -> new BlockItem(UbesDelightBlocksImpl.LEAF_FEAST_ENSAYMADA_UBE.get(), noStack()));
+            () -> new BlockItem(UbesDelightBlocksImpl.LEAF_FEAST_ENSAYMADA_UBE.get(), leafFeast()));
+    public static final Supplier<Item> LEAF_FEAST_ENSAYMADA_UBE_HALF = registerItemWithTab(UbesDelightItems.LEAF_FEAST_ENSAYMADA_UBE_HALF,
+            () -> new BreadLeafFeastItem(UbesDelightBlocksImpl.LEAF_FEAST_ENSAYMADA_UBE.get(), halfLeafFeast().component(UbesDelightDataComponentTypesImpl.SIMPLE_LEAF_FEAST_SERVINGS, 3)));
+
     public static final Supplier<Item> LEAF_FEAST_PANDESAL = registerItemWithTab(UbesDelightItems.LEAF_FEAST_PANDESAL,
-            () -> new BlockItem(UbesDelightBlocksImpl.LEAF_FEAST_PANDESAL.get(), noStack()));
+            () -> new BlockItem(UbesDelightBlocksImpl.LEAF_FEAST_PANDESAL.get(), leafFeast()));
+    public static final Supplier<Item> LEAF_FEAST_PANDESAL_HALF = registerItemWithTab(UbesDelightItems.LEAF_FEAST_PANDESAL_HALF,
+            () -> new BreadLeafFeastItem(UbesDelightBlocksImpl.LEAF_FEAST_PANDESAL.get(), halfLeafFeast().component(UbesDelightDataComponentTypesImpl.SIMPLE_LEAF_FEAST_SERVINGS, 3)));
+
     public static final Supplier<Item> LEAF_FEAST_PANDESAL_UBE = registerItemWithTab(UbesDelightItems.LEAF_FEAST_PANDESAL_UBE,
-            () -> new BlockItem(UbesDelightBlocksImpl.LEAF_FEAST_PANDESAL_UBE.get(), noStack()));
+            () -> new BlockItem(UbesDelightBlocksImpl.LEAF_FEAST_PANDESAL_UBE.get(), leafFeast()));
+    public static final Supplier<Item> LEAF_FEAST_PANDESAL_UBE_HALF = registerItemWithTab(UbesDelightItems.LEAF_FEAST_PANDESAL_UBE_HALF,
+            () -> new BreadLeafFeastItem(UbesDelightBlocksImpl.LEAF_FEAST_PANDESAL_UBE.get(), halfLeafFeast().component(UbesDelightDataComponentTypesImpl.SIMPLE_LEAF_FEAST_SERVINGS, 3)));
+
     public static final Supplier<Item> LEAF_FEAST_HOPIA_MUNGGO = registerItemWithTab(UbesDelightItems.LEAF_FEAST_HOPIA_MUNGGO,
-            () -> new BlockItem(UbesDelightBlocksImpl.LEAF_FEAST_HOPIA_MUNGGO.get(), noStack()));
+            () -> new BlockItem(UbesDelightBlocksImpl.LEAF_FEAST_HOPIA_MUNGGO.get(), leafFeast()));
+    public static final Supplier<Item> LEAF_FEAST_HOPIA_MUNGGO_HALF = registerItemWithTab(UbesDelightItems.LEAF_FEAST_HOPIA_MUNGGO_HALF,
+            () -> new BreadLeafFeastItem(UbesDelightBlocksImpl.LEAF_FEAST_HOPIA_MUNGGO.get(), halfLeafFeast().component(UbesDelightDataComponentTypesImpl.SIMPLE_LEAF_FEAST_SERVINGS, 3)));
+
     public static final Supplier<Item> LEAF_FEAST_HOPIA_UBE = registerItemWithTab(UbesDelightItems.LEAF_FEAST_HOPIA_UBE,
-            () -> new BlockItem(UbesDelightBlocksImpl.LEAF_FEAST_HOPIA_UBE.get(), noStack()));
+            () -> new BlockItem(UbesDelightBlocksImpl.LEAF_FEAST_HOPIA_UBE.get(), leafFeast()));
+    public static final Supplier<Item> LEAF_FEAST_HOPIA_UBE_HALF = registerItemWithTab(UbesDelightItems.LEAF_FEAST_HOPIA_UBE_HALF,
+            () -> new BreadLeafFeastItem(UbesDelightBlocksImpl.LEAF_FEAST_HOPIA_UBE.get(), halfLeafFeast().component(UbesDelightDataComponentTypesImpl.SIMPLE_LEAF_FEAST_SERVINGS, 3)));
+
     public static final Supplier<Item> POISONOUS_UBE = registerItemWithTab(UbesDelightItems.POISONOUS_UBE,
             () -> new Item(basicItem()));
     public static final Supplier<Item> UBE = registerItemWithTab(UbesDelightItems.UBE,