+++ changelog.md
@@ -21,10 +21,18 @@ NeoForge and Fabric.
   - Special Variants
     - Created by right-clicking a leaf feast with a valid item
       - Valid Items: Lumpia, Ensaymada, Ube Ensaymada, Pandesal, Ube Pandesla, Hopia Munggo, Hopia Ube
+      - Can also be crafted
+        - Lumpia Leaf Feast Example
+          - 1x Leaf Feast, 6x Lumpia (Shaped Crafting) -> 1x Lumpia Leaf Feast
+          - 3x Leaves, 6x Lumpia (Shaped Crafting) -> 1x Lumpia Leaf Feast
+          - 1x Leaf Feast, 3x Lumpia (Shapeless Crafting) -> 1x Half Lumpia Leaf Feast
+          - 3x Leaves, 3x Lumpia (Shaped Crafting) -> 1x Half Lumpia Leaf Feast
+      - Leaf Feasts stack to 8, Half Leaf Feasts stack to 16(also Leaf Feasts with 3 servings)
     - Can be created and refilled by dispenser's
     - Special variants can be crafted and placed with 6 servings
 - New Rolling Pin Tiers
   - Iron, Gold, Diamond, Netherite
+  - Slightly reduce attack speed
   - Better Combat compatibility
 - Ube Milk Tea Block
   - Ube Milk Tea item can now be placed similar to the Halo Halo item and upgraded block